// src/app/models/task.model.ts
export interface Task {
  dia: string;
  periodo: string;
  horas: number;
  descricao: string;
  custeio: string;
  status: string;
  os: string;
  projeto: string;
  cliente: string;
  despesasAutorizadas: string;
  alimentacao: string;
  viagem: string;
  inLoco: string;
  tipoAtendimento: string;
  executavelAlterado: string;
  versao: string;
  solicitante: string;
}